Don't change the template itself, but its data subpage instead (image part)
A commit earlier today already applied this to the descriptions template,
but now the image template itself (which includes the descriptions) is also
separated such that the template description and formatting is stored in the
template itself, and the data (which changes a lot) is in a separate subpage.

The advantage is that the page history stays cleaner -- you can more easily
see how the documentation and markup changed over time.